+++ packages/account/PasskeyAccount/passkeyOwner.ts
@@ -1,0 +1,6 @@
+import type { BytesLike, HexString } from "~typing"
+
+export interface PasskeyOwner {
+  use(credentialId: string): void
+  sign(challenge: BytesLike): Promise<HexString>
+}
